added HTTP status to APIs

+++ src/main/java/com/hyu/electronicsecwebsitebe/controller/ImportsController.java
@@ -4,6 +4,8 @@ package com.hyu.electronicsecwebsitebe.controller;
 import com.hyu.electronicsecwebsitebe.model.Imports;
 import com.hyu.electronicsecwebsitebe.service.impl.ImportsServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,27 +17,46 @@ public class ImportsController {
     private ImportsServiceImpl importsService;
 
     @GetMapping("/all")
-    public List<Imports> getAllImports() {
-        return importsService.getAllImports ();
+    public ResponseEntity<List<Imports>> getAll() {
+        List<Imports> importsList = importsService.getAllImports ();
+        return ResponseEntity.ok (importsList);
     }
 
     @GetMapping("/{id}")
-    public Imports getImportsById(@PathVariable String id) {
-        return importsService.findById (id);
+    public ResponseEntity<Imports> getImportsById(@PathVariable String id) {
+        Imports imports = importsService.findById (id);
+        if (imports != null) {
+            return ResponseEntity.ok (imports);
+        } else {
+            return ResponseEntity.notFound ().build ();
+        }
     }
 
     @PostMapping("/save")
-    public Imports saveImports(@RequestBody Imports imports) {
-        return importsService.saveImports (imports);
+    public ResponseEntity<Imports> saveImports(@RequestBody Imports imports) {
+        if (importsService.existsById (imports.getId ())) {
+            return ResponseEntity.badRequest ().build ();
+        }
+        Imports savedImports = importsService.saveImports (imports);
+        return ResponseEntity.status (HttpStatus.CREATED).body (savedImports);
     }
 
-    @PutMapping("/update")
-    public Imports updateImports(@RequestBody Imports imports) {
-        return importsService.updateImports (imports);
+    @PutMapping("/update/{id}")
+    public ResponseEntity<Imports> updateImports(@PathVariable String id, @RequestBody Imports imports) {
+        if (!importsService.existsById (id)) {
+            return ResponseEntity.notFound ().build ();
+        }
+        imports.setId (id);
+        Imports updatedImports = importsService.updateImports (imports);
+        return ResponseEntity.ok (updatedImports);
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteImports(@PathVariable String id) {
+    public ResponseEntity<Void> deleteImports(@PathVariable String id) {
+        if (!importsService.existsById (id)) {
+            return ResponseEntity.notFound ().build ();
+        }
         importsService.deleteImports (id);
+        return ResponseEntity.noContent ().build ();
     }
 }
